_LEGACY_BACKUP/backend/requirements.txt
flask>=3.0.0
rapidfuzz>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pytest>=7.0.0
